Adding ConstructionWand data
This is for the [Construction Wand](https://github.com/Theta-Dev/ConstructionWand) mod.

I pulled the craft values for the wands from shovels of appropriate types, as I did the same with the salvage.

For the infinity wand (uses nether star), I used diamond values but a little higher (harder to salvage, more xp for crafting).

For the two cores, I just picked the most expensive resource in each to give salvage for (diamond block for demon, 2 gold bars for angel), and gave them appropriate salvage rates. Crafting value on them was compared to pickaxes of the appropriate material types (but scaled up massively on demon core since it uses a block of diamond and two diamond picks).